+++ handler/core/src/main/java/io/knotx/fragments/handler/action/cb/CircuitBreakerActionFactoryOptions.java
@@ -16,10 +16,13 @@
 package io.knotx.fragments.handler.action.cb;
 
 import static io.knotx.fragments.handler.api.actionlog.ActionLogLevel.ERROR;
+import static io.knotx.fragments.handler.api.domain.FragmentResult.ERROR_TRANSITION;
 
 import io.vertx.circuitbreaker.CircuitBreakerOptions;
 import io.vertx.codegen.annotations.DataObject;
 import io.vertx.core.json.JsonObject;
+import java.util.HashSet;
+import java.util.Set;
 import java.util.UUID;
 import org.apache.commons.lang3.StringUtils;
 
@@ -31,12 +34,11 @@ public class CircuitBreakerActionFactoryOptions {
    */
   private static final boolean DEFAULT_FALLBACK_ON_FAILURE = true;
 
-  private static final CircuitBreakerOptions DEFAULT_CIRCUIT_BREAKER_OPTIONS = initCircuitBreakerOptions(
-      new CircuitBreakerOptions());
-
   private String circuitBreakerName;
 
-  private CircuitBreakerOptions circuitBreakerOptions = DEFAULT_CIRCUIT_BREAKER_OPTIONS;
+  private CircuitBreakerOptions circuitBreakerOptions = new CircuitBreakerOptions();
+
+  private Set<String> errorTransitions = new HashSet<>();
 
   private String logLevel = ERROR.getLevel();
 
@@ -55,13 +57,6 @@ public class CircuitBreakerActionFactoryOptions {
    */
   public CircuitBreakerActionFactoryOptions(JsonObject json) {
     CircuitBreakerActionFactoryOptionsConverter.fromJson(json, this);
-    this.circuitBreakerOptions = initCircuitBreakerOptions(circuitBreakerOptions);
-  }
-
-  private static CircuitBreakerOptions initCircuitBreakerOptions(CircuitBreakerOptions options) {
-    CircuitBreakerOptions newOptions = new CircuitBreakerOptions(options);
-    newOptions.setFallbackOnFailure(DEFAULT_FALLBACK_ON_FAILURE);
-    return newOptions;
   }
 
   /**
@@ -96,7 +91,9 @@ public class CircuitBreakerActionFactoryOptions {
    * @return the circuit breaker configuration options
    */
   public CircuitBreakerOptions getCircuitBreakerOptions() {
-    return circuitBreakerOptions;
+    CircuitBreakerOptions result = new CircuitBreakerOptions(circuitBreakerOptions);
+    result.setFallbackOnFailure(DEFAULT_FALLBACK_ON_FAILURE);
+    return result;
   }
 
   /**
@@ -109,6 +106,26 @@ public class CircuitBreakerActionFactoryOptions {
   public CircuitBreakerActionFactoryOptions setCircuitBreakerOptions(
       CircuitBreakerOptions circuitBreakerOptions) {
     this.circuitBreakerOptions = circuitBreakerOptions;
+    return this;
+  }
+
+  /**
+   * @return transitions that mean error
+   */
+  public Set<String> getErrorTransitions() {
+    Set<String> result = new HashSet<>(errorTransitions);
+    result.add(ERROR_TRANSITION);
+    return result;
+  }
+
+  /**
+   * Sets error transitions.
+   *
+   * @param errorTransitions transitions that mean error
+   * @return the current {@link CircuitBreakerActionFactoryOptions} instance
+   */
+  public CircuitBreakerActionFactoryOptions setErrorTransitions(Set<String> errorTransitions) {
+    this.errorTransitions = errorTransitions;
     return this;
   }
 
@@ -128,5 +145,15 @@ public class CircuitBreakerActionFactoryOptions {
   public CircuitBreakerActionFactoryOptions setLogLevel(String logLevel) {
     this.logLevel = logLevel;
     return this;
+  }
+
+  @Override
+  public String toString() {
+    return "CircuitBreakerActionFactoryOptions{" +
+        "circuitBreakerName='" + circuitBreakerName + '\'' +
+        ", circuitBreakerOptions=" + circuitBreakerOptions +
+        ", errorTransitions=" + errorTransitions +
+        ", logLevel='" + logLevel + '\'' +
+        '}';
   }
 }